refactor tests add entrypoint tests
